add property splitting delimiter as a constant
